Reduce memory dependency
